development completed

+++ laravell_new/app/Http/Controllers/ListingController.php
@@ -47,6 +47,7 @@ class ListingController extends Controller
         if ($request->hasFile('logo')) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
+        $formFields['user_id'] = auth()->id();
 
         Listing::create($formFields);
 
@@ -57,6 +58,12 @@ class ListingController extends Controller
     // Update Listing Data
     public function update(Request $request, listing $listing)
     {
+
+        //Make Sure Logged In User Is Owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorised Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -79,6 +86,11 @@ class ListingController extends Controller
     // Delete Listing
     public function destroy(Listing $listing)
     {
+        //Make Sure Logged In User Is Owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorised Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing Deleted Successfully!');
     }
@@ -87,5 +99,12 @@ class ListingController extends Controller
     public function edit(Listing $listing)
     {
         return view('listings.edit', ['listing' => $listing]);
+    }
+
+    //Manage Listings
+
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
